Edit: Sceleton, Interface and first argument
The sceleton for the interface and it's first possible argument are
finished, so a users pbukey can be changed now without touching the DB
itself directly.

Checking for users existence and the presence of options/arguments is
done beforehand and will throw out when nothing useful got delivered to
the script.